🚚 Move UserInfos in own folder
