+++ src/pages/UserInfos/UserInfos.svelte
@@ -1,1 +1,5 @@
 <h2>User infos</h2>
+
+<form>
+  
+</form>
